requirements.txt
mysql-connector-python
tabulate
matplotlib
Pillow
Flask
Flask-CORS
python-dotenv
requests
APScheduler
